requirements.txt
# IT Diligence Analysis Agent
# Requirements for deployment

# Anthropic API
anthropic>=0.39.0

# PDF Processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# Environment
python-dotenv>=1.0.0

# Data handling
pydantic>=2.0.0

# Async operations
aiofiles>=23.0.0

# Web UI - Streamlit
streamlit>=1.30.0

# Web UI - Flask (results browser)
flask>=3.0.0
gunicorn>=21.0.0

# Excel export
openpyxl>=3.1.0
